Change in simulate from items
I was not able to simulate data from an items object. I am not a 100% sure that my fix is entirely correct but it seems that the ifelse statement was unintentionally filling out the length.out argument in stead of the by argument (and that the greater than sign should have been a less than sign). By tracing the function with this alteration i was able to run the function with no problems. 

Let me know if i have completely misunderstood the purpose of the ifelse clause and i should look closer for a proper fix :)